BUG: Restore backwards compatibility
The commit c8b1b5f715bc9e7419a38251d5210702b967c144 was
a partial solution to allowing use of image spacing
consistently.

Previously the m_UseImageSpacing was defaulted to
true, but ignored in the case where "SetInputImage",
was the only function that changed changed the default
state of the function.

It is not possible to retain perfect backwards
compatibility.  The behavior of the previous
implementation changed based the order and
values of changing the state.

Previously, there was also no mechanism
for manually setting the value of UseImageSpacing
from the end-user.

This commit adds the ability to change behavior by
defaulting m_UseImageSpacing to true, while also
providing end-users the abiltiy to now properly request
that UseImageSpacing be explicitly turned On or Off.

Change-Id: Idb8b02dedce2dc4a220539a47d0762810651270c