PERF: NeighborhoodOperatorImageFunction avoids copy ConstNeighborhoodIterator
Replaced default-construction + assignment of 'bit' (a
ConstNeighborhoodIterator) by a construction of 'bit' that
directly initializes its radius, image, and region.

Also added assert (debug-only) that image pointer is valid.

This may look like a stylish commit, but it does significantly improve the
performance of NeighborhoodOperatorImageFunction::EvaluateAtIndex, and
thereby of GaussianDerivativeImageFunction and
HoughTransform2DCirclesImageFilter. In some cases, the run-time duration is
reduced by more than 30%.

Change-Id: I91fd8edc146af6a74f003fa62effac699bb71714