ENH: update C+11 headers and try compile for headers
The try compiles for "stddef.h" and "sys/types.h" are not used in ITK
so they were removed.

The try compile for stdint.h has been removed for the expected C+11
standard cstdint.

fenv.h remains used in itkFloatingPointExceptions.cxx because the
usage is low level and system specific as opposed to complying with a
standard usage.

Change-Id: I9e176ae02de2eb9c7ec59bcb66ed884f16c61dab