ENH: Adding user set min and max values for noise.
The original filter settings used the min and max values based on the
pixel type of the output image. The current addition allows the user to
set the min and max values with the defaults consistent with the
previous settings. This is desirable when dealing with images such as
CT which have 16 bits signed integer, but only use 12 bits to
represent values.

Change-Id: I0d3683ae8ce6dcaf32027e0caa23d7290afe0d41