BUG: Preserve option to use a "label value" mask and boolean mask
There are multiple ways to define where a mask image is true in
ITK. Previously, when ITK_LEGACY_REMOVE was enable, the new behavior
was set and simultaneously the option for legacy behavior was
removed. This patch changes the default behavior to the new "boolean
mask" ( an image mask where pixel_value!=0 ).

Change-Id: Ic772d03bb5db5fb48906d0f92347eaad03263951