PERF: Removed blurring from GaussianDerivativeImageFunction
Removed Gaussian blurring from GaussianDerivativeImageFunction::EvaluateAtIndex
as the blurring usually took a lot of time (possibly 50 % or more), whereas its
result, as stored in 'value', would be multiplied by 'centerval', which would
always be zero.

It appears a bug that the blurring had zero effect, and this bug appeared to
be there already with the very first commit, 2003-05-29, SHA-1:
27973617ab1b7164ae98407bd3aefffa62c18ab0 (when the code was still located at
Code/Common/itkGaussianDerivativeImageFunction.txx).

Removed creation of Gaussian blurring kernel from RecomputeGaussianKernel(),
as it is no longer used by now. Also removed GaussianFunctionType,
GaussianFunctionPointer, RecomputeContinuousGaussianKernel(const double *),
m_ContinuousOperatorArray, m_GaussianFunction, as they were not used (anymore).

Note: A more effective blurring is still to be implemented, to be added to
a future version of ITK.

See also the discussion at
https://discourse.itk.org/t/hough-transform-2d-circles-image-filter-getcircles-patch/350/39

Change-Id: I51b0976c67d5be320c009b31d98e6f3f014fc1af