COMP: Suppress invalid warning
JIRA ITK-3621 issue resolved.

Newer version of clang warn when a static constexpr template variable
is declared in a different compilation unit than the defintion.

If this split declaration/definition is done incorrectly, clang
issues a warning every time the delaration is encountered, and
a linker error when linkage occurs.

If this split delcaration/definition is done correctly, clang
annoyingly issues a compiler warning every time the declaration is
encountered, but then links without issue.

Change-Id: I678bccfbbd0ca979be65af8624c196df9eca9101