STYLE: Use copy-swap idiom for SmartPointer
Improve the implementation to follow best practices with the copy-swap
idiom for the regular and the move assignment operators. This approach
reduces code, improve maintainability, and is less error prone.

A subtle implementation note is that the assignment operator no longer
takes a reference to the operand, instead it is implicitly copied.

Change-Id: I9a63909daea3717dd91e52664ac3cc64c440aa8e