ENH: Remove TEMPLATED_FRIEND macro and try compile
Modern C++11 compiler support a standard complaint declaration of
templates friend functions. Additionally, this macro is not used in
ITK.

The macro ITK_FRIEND_TEMPLATE_FUNCTION_ARGUMENT has been deprecated an
error is produced when ITK_LEGACY_REMOVE is enabled.

The following CMake try_compiles defines are removed:
 ITK_SUPPORTS_TEMPLATED_FRIEND_FUNCTION_WITH_TEMPLATE_ARGUMENTS
 ITK_SUPPORTS_TEMPLATED_FRIEND_FUNCTION_WITH_EMPTY_BRACKETS
 ITK_SUPPORTS_TEMPLATED_FRIEND_FUNCTION_WITH_NULL_STRING
These are removed without usage errors because they are not considered
part of the public interface of ITK.

Change-Id: I5bf5e3d5eff120dde674bd9c90873b8b4b529b31