DOC: Improve exception message to precise what image sizes can be processed
When using Vnl FFT implementation, input images must have a size that
is a combination of 2,3, and 5 as prime factors. The previous wording
in the doxygen documentation and exception message was ambiguous and
could let the developer and user think that the size along each dimension
had to be a multiple of 2,3, and 5, but was not restricted to this
list (i.e. could have additional prime factors).

Change-Id: Ia49acf053dd82c89c7257b02600ea66f1c28e478