BUG: Prevent gdcm "missing implementation" error on macOS
This commit prevents Slicer from crashing with a "missing implementation"
error reported by function System::GetCurrentProcessFileName.

It comments the call to "LoadDefaultPaths" that was calling the function
"GetCurrentProcessFileName" when the GDCM library was loaded.

As already mentioned in commits 857fd54 (COMP: Set GDCM_USE_COREFOUNDATION_LIBRARY
off), b388a3b (COMP: Disable CoreFoundation calls) and b853e6c (ENH: Supression of
GDCM's linking to CoreFoundation), the function "GetCurrentProcessFileName"
is not needed because (1) the gdcm executable are not built and (2) the
resource paths added in "LoadDefaultPaths()" are not used.

See https://discourse.slicer.org/t/launch-error-message-in-qt5-debug-build/2039

Change-Id: I23bdbbaeb6420cd122ca1b6da82ef774dd79a91a