BUG: Refactor itkFloatingPointExceptions to fix macOS support
This commit addresses an issue occurring on newer macOS where the functions
"feenableexcept" and "fedisablexcept" are not available.

The message "FloatingPointExceptions are not supported on this platform."
was displayed when executing code enabling floating point exceptions. This
happened because the logic checking the support on different platform was
incorrect.

Due to the complexity of the implementation, this commit (1) breaks down
the implementation in few files, (2) add comments and (3) differentiate the
following cases for implementing feenableexcept and fedisableexcept functions:
* the function should be a no-op (e.g emscripten, solaris)
* has builtin  feenableexcept and fedisableexcept (e.g linux)
* has builtin fegetenv and fesetenv allowing to emulate
  feenableexcept and fedisableexcept (e.g macOs)

It also extend the API of itkFloatingPointExceptions with a new function
"HasFloatingPointExceptionsSupport()" allowing to check if its supported
on the current platform. This allows to simplify code in itkNrrdImageIO and
itkTestDriver.

It also generalizes the emulation of feenableexcept and fedisableexcept using
egetenv and fesetenv so that it can works with "fenv_t" struct having
different "control" member name. For example, on macOS it is "__control",
on Linux with GlibC it is "__control_word", and on Alpine where musl is
used, it is "__cw".

Finally, it also updates the implementation of FloatingPointExceptions::Enable(),
FloatingPointExceptions::Disable() and the signal handler on unix by removing
the use of FPE_* flag with feenableexcept, fedisableexcept and feclearexcept.
Indeed, FPE_* flags are values associated with the signal, They have no meaning
in the context of fenv exceptions.

Some useful references:
* http://lists.lugod.org/pipermail/vox-tech/2002-March/001322.html
* https://linux.die.net/man/3/fegetenv

Change-Id: Ia0f43dad2c2b71c575b6c93be2602994506a89da