ENH: Import the ITKBridgeNumPy module
This is imported from https://github.com/InsightSoftwareConsortium/ITKBridgeNumPy.git
@e74bdafc2824c7f63fdc7c0e36ae843db0c92393

This Remote Module is mature and is always used when building the Python
wrapping. Import it into the repository to make it easier for packagers,
so they do not have to package it separately or worry about pulling it
down at CMake configure time.

Change-Id: Ic338eebb6502c8876fb8c0707e1a2e0623223043