ENH: Make const operator[] conform to standards
The itkSize.h, itkIndex.h, and itkOffset.h provided
subtle different interface for the const version of
operator[] const.

git grep "operator\[\] *( *.*int"

exposes that only these 3 classes have a different
interface than the rest of the ITK toolkit.  Additionally
these three classes diverge from the interfaces of the
stl, in particular std::array.

This exposed a bug in GCC 5.4.0 where a reference
to a local component of a temporary variable was not
being set properly.

Change-Id: Iaa17d49f87ac06582b734e7b2d15c2022db882ab