BUG: Update NULL pointer patch with final fixes
Provide backwards compatibilty to construct the itkSmartPointer with
NULL.  External applications may have initialization with NULL.

This interface is intended to be a stopgap measure to ease
transition to ITKv5.  The new interfaces are purposefully
placed in the immediately deprecated interface.

Constructor signature needs to be non-constexpr, use an
assert to test for values, and be non-explicit in order
for examples to pass.

Added GTest to demonstrate desired behavior.

Change-Id: Ibbc45a94846cea58077b4594d4d34dceb7b725d0