ENH: Address incorrect types with neighborhood iterator base class
Propagating the correct OutputImage type removes the need for the
brute force reinterpret_cast.

The type uses for the output image type was not considered when
defining the base class for the boundary condition  in the const
neighborhood iterator. Proper typedefs were added to the
ImageBoundaryConsition class so they could be utilized by the
neighborhood iterator.

Adding support to NeighborhoodAccessorFunction for BoundaryConditions
which have a different output type than the input.

Co-authored: Matt McCormick <matt.mccormick@kitware.com> for tracking
down the cause.

Change-Id: Ib6b75ed4628c5cc8960e28ad9440460754d97093