PERF: Implement copy on write for MetaDataDictionary
The default behavior when reading a Image in ITK is to create 3 copies
of the MetaDataDictionary for the ImageIO, ImageFileReader, and Image
object. These are usually not modified, not shared and created by a
deep copy. The internal std::map is shared via a std::share_ptr with a
reference count. If the dictionaries are modified a deep copy is
performed.