BUG: Add function to not wait for threads when destroying thread pool
In certain cases, on Windows, the application using ITK will hang when trying
to wait for the threads created by the threadpool. These threads have already
been destroyed and the system is waiting for nothing. This problem has been
addressed when ITK is built dynamically [1] but ITK can be built statically
and integrated in a dynamic library (such as in ITK Python). In this case,
we can use this programmatic method to not wait for threads to avoid
the process hanging upon exit.

[1] 037e44829fd5b69ce31515fb04992c054fe3f282

Change-Id: I91787aed00642f6d789a1e05d9e317f5ee8b9510