PERF: Remove iterators stored as member variables
The methods Begin and End now return an iterator constructed on the
fly, instead of a stored iterator stored as a member variable. New
constructors were added for efficient construction of the
iterators. This removed the need to maintain consistency of the
iterators and duplication in the two classes.

A 15% speed up on the ITKPerformanceBenchmark MohphologicalWatershed
has occurred with this improvement.

Change-Id: I7861ac428b694a50be4af237384b1c61087d2790