ENH: Adding Unary and Binary GeneratorImageFilter
The current ITK functor based filters are problematic because they
create code blot. The base classes UnaryFunctorFilter,
BinaryFunctorFilter etc, take a template parameter of the functor. So
these "base" classes have different instantiates for every derived
class. The generate code of these base classes is not reused!

To address this, we now introduce a base class which does not have a
template parameter of the Functor Type. Instead there is a templated
"SetFunctor" method which generates the appropriate
ThreadedGenerateData method with a copy of the functor.

Change-Id: Iaf47f4b34413962cbfeb6f54deb57924c5c47aa9