STYLE: increasing initial size of the debug window
The initial size of 512 by 512 pixels is not enough to properly show
majority of text which is displayed in it. The text is usually cut off.
This is especially pronounced if ITK is located in a long path.
A high-DPI screen makes each letter wide in absolute pixel count.
Screen resolution statistics support moving to a larger initial window:
http://gs.statcounter.com/screen-resolution-stats/desktop/worldwide
It is also expected that software developers, as professional computer
users, have larger than average screens.

Change-Id: Icd46a60922a48c9f345e8606db019b2274d7e895