STYLE: Removed '&' from pixelPointer param of NeighborhoodAccessorFunctor::Set
The first parameter of the Set(pixelPointer, p) member function of both
NeighborhoodAccessorFunctor and VectorImageNeighborhoodAccessorFunctor
was declared as non-const reference-to-pointer, 'InternalPixelType * &',
suggesting that they might change the actual pointer, to start pointing
somewhere else. But none of them does actually change the pointer. As
is now checked, during compilation, by declaring the pixelPointer
parameter as a 'const' pointer, instead of a non-const
reference-to-pointer.

As this commit removed the reference, it removes an extra indirection
during data access, which might improve the performance. It certainly
won't harm!

Change-Id: I0d6292a9061a3debc9517ff66d97fad680345b9e