ENH: Improve numerics of resampling for linear transform
The method of computing the sample points for linear transforms has
been improved to remove iterative accumulation error, and enable the
computation to be independent of the processing region. This is done
by using linear interpolation between the staring and ending
continuous index instead of accumulating a "delta" continuous index.

RecursiveMultiResolutionPyramidImageFilter test is updated for
streaming, because internally the pyramid filter is breaking up the
image, changing the largest possible region, which violates the
assumptions for reproducible computations for this linear
resampling. The difference in values is less than 2 ulps.

Replaced the baseline test image for ResampleImageFilter3Test1. The
new results are different by 3 pixels on the first scan line.

Added an additional base line for the ImageRegistration10Test, this
test resamples a label image with the nearest neighborhood
interpolator. The difference with the prior results was 33 pixels at
the edges of the labels.

Change-Id: Ib4c1018f8b0b77ca730144af4f17999e04bb3fa8