PERF: Removed virtual ~NeighborhoodAccessorFunctor() and made class final
It appears that NeighborhoodAccessorFunctor is not being used as a
base class, because it can be declared 'final' without compilation
errors. Clearly, if it is not used as base class, it does not need
to have a virtual destructor.

By removing the virtual destructor, a significant performance gain
can be achieved, for both the classic ITK NeighborhoodIterator
classes and the new ShapedImageNeighborhoodRange. Using MSVC2015,
~7% reduction of the run-time duration of GaussianDerivativeImageFunction
(as used by HoughTransform2DCirclesImageFilter::Update()) was observed.

This commit simply removes both the empty destructor and the empty
default-constructor, following "the rule of zero":
https://github.com/isocpp/CppCoreGuidelines/blob/master/CppCoreGuidelines.md#c20-if-you-can-avoid-defining-default-operations-do

It also adjusts the ShapedImageNeighborhoodRange constructor that
copies the NeighborhoodAccessorFunctor from the image, in order to
avoid AppleClang 6.0.0.6000056 compile errors ("excess elements in
struct initializer").

Change-Id: Ibda42f4dab43d31bc50bc61c02dcb38428f93f8b