ENH: Allow for itk::ImageSource wrapping override
Find itkImage template if empty when defining itkImageSource typemaps
for python wrapping. An empty value for the template means that the
corresponding type was not requested by the ITK_WRAP_* options.
It can happen if an external module wants to add a specific type for
ImageSource that was not requested in a first place in ITK.
e.g. calling itk_wrap_template() on ImageSource for type double without
requesting ITK_WRAP_double.

Change-Id: I1db4b889d751a0ce30a6087fb44210a92a9ff3d6