ENH: Add frequency iterators and band pass filter.
Initially developed for ITKIsotropicWavelets module [1].
The associated Insight Journal article contains a descriptions of these
iterators [2].
The frequency iterators are ImageIterators with GetFrequency and
GetFrequencyBins methods to abstract and ease development of filters in
the frequency domain.

Images in the frequency domain have layouts that depend on the
algorithm that generates them, for example after a FFT, the zero
frequency is at origin, max positive frequency is at index size/2 - 1.
The min negative frequencies is size/2, and the least negative
frequency is at index: size - 1.
This layout change slightly if image is even or odd in size.
(Nyquist bands appear).

The best example is
itkFrequencyFFTLayoutImageRegionConstIteratorWithIndex.h
that contains the layout after applying the standard
Fast Fourier Transform to an image.
Also the Shifted version, that cover the common case
of shifting the zero frequency to the center of the image.

These iterators try to abstract this complexity from the developer, for
them to only manipulate frequency values. The best example is the
FrequencyBandFilter, that acts as BandPass or BandStop filter given two
boundary frequencies.

The base classes of this hierarchy of iterators handle
the case where images are taken
in frequency space (scattering experiments for example),
and they are just regular ImageIterator with the GetFrequency
functions.

Note: The frequency iterator for hermitian images is not included
(FFT of real images can be chopped by half because they contain
redundant information)

Also at the time of developing I triggered this question in the
mail-list about the lack of meaning of image metadata after performing
a FFT: [3]

[1]: https://github.com/phcerdan/ITKIsotropicWavelets/
[2]: http://www.insight-journal.org/browse/publication/986
[3]: https://public.kitware.com/pipermail/community/2017-May/013055.html

Change-Id: Ie29319eb66cd0dd9f95545396faa7df9226a9364