ENH: Add stateless viewer ViewImage with Python wrap.
Use a simple function (transformed in a class with a static function
for wrapping reasons [1] ) to visualize 2D AND 3D images.

Added to ITKVtkGlue.
It can be used in python with itkViewImage[ImageType].View(inputImage)
This should provide a method to visualize images in itk without need of
third party software.

I use this function every time, and could be useful for others.
It also can be used from the console, if tests are enabled:

/path/ITK-build/bin/ITKVtkGlueTestDriver runViewImage /path/image.xxx

[1]: https://discourse.itk.org/t/python-wrapping-error-cannot-define-or-redeclare-x-here-because-namespace-wrappers-does-not-enclose-namespace-itk/886/9

Change-Id: Ieb16fdc846134b8e962b0d78b424809b2797c9ae