ENH: Added class to create offsets for a hyperrectangular neighborhood shape
Added HyperrectangularImageNeighborhoodShape<ImageDimension> class,
to ease creating a sequence of offsets for a hyperrectangular shape,
to be used with the new (ITK 5.0) ShapedImageNeighborhoodRange, as
well as with the original itk::ShapedNeighborhoodIterator.

Aims to suggest (rather than to impose) a general design for any
custom neighborhood shape class:

  class CustomImageNeighborhoodShape
  {
  public:
    static constexpr unsigned int ImageDimension = N;
    constexpr size_t GetNumberOfOffsets() const noexcept;
    void FillOffsets(Offset<N>* offsets) const noexcept;
  };

Adapted "itkImageNeighborhoodOffsets.h" to start using
HyperrectangularImageNeighborhoodShape, and added generic
convenience function, GenerateImageNeighborhoodOffsets(shape).
Removed FillHyperrectangularImageNeighborhoodOffsets, which is
redundant now, and was never used before anyway.

Change-Id: Ie96ae9e894ab3a6b0f269379394084673fcf8754