BUG: Prevent concurrent read/write in output image
Coarsely fix threading issue with IsoContourDistanceImageFilter,
utilized by the ApproximateSignedDistance, and NarrowBand LevelSet
framework. This patch address a periodic failing test on the ITK
dashboard: itkApproximateSignedDistanceMapImageFilterTest1. And likely
address NarrowBand predictability issues previously encountered.

This patch oafishly utilizes a mutex lock when reading and writing to
the output image. Given, that this filter only updates the values
around the iso-contour the locking does not occur for every pixel.

The algorithm should be re-witten with out the concurrent read/write
problems. Potential solutions include, having each thread allocate a separate
output buffer, then merging. Another is to crop the regions and handle
the overlapping boundaries separately. But this would heavily depend
on how the output regions are split, and there is a trend to have less
control over the splitting methods.

Change-Id: I22d74a5e948d203f71a71df60353e3852ec5dbfb