PERF: less strict heuristic for valid points check
The previous check was overly strict.   For example, one may be using a small mask (e.g. hippocampus) to focus registration within a much larger image.   The registration result can be valid in this case despite a relatively small number of points used.  Another alternative to this request would be to remove this check entirely.