BUG: Fix unused variable CDash warnings.
Fix the unused variable CDash warnings reported in:
https://open.cdash.org/viewBuildError.php?type=1&buildid=5401884
and issued by the builds triggered in the following gerrit topic:
http://review.source.kitware.com/#/c/23434/

Also, take advantage to remove the unused `NumberOfRows` `typename` in the
same file. Moreover, the new multi-threading framework does not require
the filters to declare and update variables to report their progress.