PERF: improve DCMTKFileReader::CanReadFile perf
Using DcmMetaInfo avoids loading full dataset into memory, which
can be very expensive for large files.
For example, in Slicer, DCMTK is one of the first ITK readers tested.
Without this change, attempting to read a multi-gigabyte NIfTI file
will hang for many minutes in the ::CanReadFile code path. With this
change, the test path completes in under 50ms.