ENH: Create TestKernel library
Create a library for the TesKernel code. Specifically, the code
that explicitly registers the IO factories and performs the image
comparisons is moved from header files to compiled code. Th
removes the public dependencies for TestKernel module on the many of
the ImageIO and their third party libraries, this reduces the includes
paths and reduces compilation time by not recompiling this code for
every test driver. Additionally, the image comparison code is now
compiled into the library.

Change-Id: Icbab45175993496d96c117dc0a61d227fad560ff