STYLE: Removed HoughTransform2DCircles default for TRadiusPixelType
ITK 4.13 introduced TRadiusPixelType as an extra template parameter
of HoughTransform2DCirclesImageFilter. It had a default argument,
TRadiusPixelType = TOutputPixelType, for ITK4 backward compatibility.
Unfortunately, this default argument is often not the right choice.
For OutputPixelType, an unsigned integer is often preferable, whereas
for RadiusPixelType, a floating point type is often a better choice.

This commit removes the default argument from TRadiusPixelType,
hoping to trigger ITK users to make the most appropriate choice.

Change-Id: I34d8b543815192700ad01c04d9c37df256ae2639