ENH: TileImageFilter learns to use VectorImages
Update the TileImageFilter to support VectorImages which have variable
length vectors. The default pixel value is updated so that it's still
ZeroValue for scalar and fixed length vectors but is zero length for
variable length vectors. If the default pixel value is zero length,
then the value is automatically resized to the proper length and
filled with zeros. The internal image and filter types have learned to
deal with VectorImages to obtain the proper types.

Additionally, the number of components must be propagated from the
input to the output. And all inputs must have the same number of
components.

Adding GTest for TileImageFilter.

Change-Id: I84be6df70db0d15397db1c44c2c88556158e4ed4