WIP: Add itkLoad method to IO plugin modules
ITK has the capability to search in a directory for shared libraries
which contain an exported "itkLoad" method. This method returns a
factory, which is registered. Thus enabling the expansion of the
supported IOs at run-time rather than compile time.

Use cases include making remote or external modules, IO plugins so
that they could be utilized by binaries unaware to the option during
initial compilation.

Change-Id: I33e7537b4e61b3b5f6968955528ce67990db54b1