BUG: Explicitly set sampled pointset in virtual domain
For using a sampled point set for a metric in the
ImageRegistrationMethodv4, the point set was generated by iterating
over the virtual domain. This matches the way the dense sampling is
done. However, this is set to the metrics as the *fixed* point set,
which the metric then transform from the fixed image domain to the
virtual image domain, essentially applying the initial fixed image
transform twice.

Now a SetVirtualSampledPointSet method is exposed and a new state
member variable UseVirtualSampledPointSet.

The widely used UseFixedSampledPointSet has been renamed to just
UseSampledPointSet which indicates either the FixedSampledPointSet or
the VirtualSampledPointSet should be used based on the value of
UseVirtualSampledPointSet.

Change-Id: I232a3d22fd752ef1ee9f9afb539542e383ac1b97