PERF: Memory alloc reduction in ComputeJacobianWithRespect
Dynamic memory allocation on a per sample or per pixel basis is a
major bottle neck for performance especially multiprocessing
scalability. The
CompositeTransform::ComputeJacobianWithRespectToParameters is called
on a per sample basis.

This patch special cases the common case of only one transform in the
composite transform and eliminates any dynamic allocation. Also
multiplication of sub-matrices has be manually done in a for loop to
reduce 3 memory allocations down to 1 in that code block.

Change-Id: Ifcdb48d78bea20303a1dca616a92ec386464ebb6