ENH: Provide explicitly instantiated vnl_svd_fixed
More dimensions are needed to support the usage of the fixed svd algorithm
for compile time known matrix sizes to provide performance improvements
in jacobian calculations.

A variety of vnl classes are explicitly instantiated in the
VNLInstatiation library.

This patch/approach provides support of fixed
svd implementations that are explicitly instantiated for all
types/dimensions of the instantiated itk::Transform.

Change-Id: I9cd806da3c96646aab66692920b2ac3f4446e3d3