PERF: Change the purpose of the cache jacobian argument
In the CompositeTransform, reuse the cache parameter for the transform
to optimize's Jabobian with respect to parameters. For the common case
where only one transform is being optimized the size of the matrix
will remain the same, thus eliminating the last dynamic memory
allocation in this method. This cache jacobian argument may not be
used if there is only one transform the the composite transform, so
pre-allocation of the jacobian has been removed.

Change-Id: I9500fb8cd42941c95b645e608166063b4cf6b32b