ENH: Update itkNaryFunctorImageFilter
There is a need for a itk::NaryFunctorImageFilter because
a different number of input images may depend on the context.

Specifically, one may have 1 to 3 input images. The filter
itk::NaryFunctorImageFilter has a required number of inputs equal to 2

Above all, the method
DynamicThreadedGenerateData with the following check:
numberOfValidInputImages == 0 and I am wondering why it is not
numberOfValidInputImages < 1.

Therefore the suggested change is to set the number of required inputs
to 1.