STYLE: Consistently name for ITK conventions
We already have a name for this behavior in ITK, and we should maintain
things that do similar things having similar things. It becomes a mess
when the same thing is named differently, and very burdensome for using
trying to find similar things. Please follow the existing convention of
" ZeroFluxNeumann" in stead of "BorderReplication" ( that name is from a
different toolkit not ITK, use the ITK name for the behavior).

It is important to understand that consistency in the ITK toolkit is
more important than trying to match external tools naming conventions.
The consistency impacts how well people can understand and use the ITK
toolkit.

Change-Id: I787d5c842cdc845fb4b34614321544276096cd19