FIX: Large erosions were producing blank images
Edge effects in morphology operations are typically handled by
including an extreme boundary value (max of pix type
for erosion, min for dilation), that stops the boundary
having any influence on the operation. The value was being
set using the relevant numeric traits of the input image type.
This is incorrect for the label morphology filters as the
work is actually done on the distance image. This patch
sets extreme values based on the pixel type of the internal
distance image.

Some unused variables have been removed.