STYLE: Prefer error checked std::stoi over atoi
The atoi function does not provide mechanisms for distinguishing between
'0' and the error condion where the input can not be converted.

std::stoi provides exception handling and detects when an invalid
string attempts to be converted to an integer.

atoi()
  Con: No error handling.
  Con: Handle neither hexadecimal nor octal.

The use of atoi in code can cause it to be subtly broken.
atoi makes two very big assumptions indeed:
  The string represents an integer/floating point value.
  The integer can fit into an int.

Change-Id: Icc43ff8e70fdd110066278d8bb3b4b284d257e3d