STYLE: Prefer error checked std::stod over atof
The atof function does not provide mechanisms for distinguishing between
'0.0' and the error condion where the input can not be converted.

std::stod provides exception handling and detects when an invalid
string attempts to be converted to an double precision number.

atof()
  Con: No error handling.

The use of atof in code can cause it to be subtly broken.
atof makes a very big assumptions indeed:
  The string represents a floating point value.

Change-Id: I4360c88f9968d0e4fefd22075904c1e7ac349f2b