ENH: Add ImageIO methods to check name for file extensions
Create new ImageIO methods which utilize the internal file extensions
lists to check a filename for a supported extension. The ImageIOs are
then refactored to use these new methods.

Some ImageIO specified lower and upper case extensions e.g. ".bmp",
".BMP". For this case case insensitive comparison of the extensions
was done. For cases where just the lower case extensions was listed,
case insensitive comparison of the extensions was chosen.

Change-Id: I4e151365e4d40f755b37412d8ff047bd9ce04fdd