BUG: Disable using mkl_tbb_thread
MKL can use either the `mkl_sequential` library, or the `mkl_tbb_thread` library.
When using the `mkl_tbb_thread` library, it requires TBB, which necessitate more
CMake code to be configured correctly and passed to projects that use MKL. Support
for this is currently broken in ITK and therefore was disabled. A follow-up commit
could solve the linking problem, but this commit is solely to remove what is broken.

Change-Id: I94f1b76c7956f08a666e3f005d5719163c83b2aa