STYLE: Replaced "Hyperrect..." by "Rect..." in class + function names
Renamed the shape class 'HyperrectangularImageNeighborhoodShape' to
'RectangularImageNeighborhoodShape', and renamed the function
'GenerateHyperrectangularImageNeighborhoodOffsets' to
'GenerateRectangularImageNeighborhoodOffsets'.

The 'hyper' adjective is not really necessary, as the template
parameter 'VImageDimension' already indicates that the rectangle may
be generalized for higher dimensionalities. Moreover, the template
also supports 1D, while 'HyperrectangularImageNeighborhoodShape' might
(incorrectly) suggest that the dimensionality should be > 2.

Change-Id: I750fe282282eba43441033f0e16cff9312e28915