ENH: Added policy for constant NeighborhoodRange values outside image
Added ConstantBoundaryImageNeighborhoodPixelAccessPolicy - a policy to get a specified constant value from ShapedImageNeighborhoodRange when a pixel value outside the image is queried. Equivalent to itk::ConstantBoundaryCondition with itk::ConstNeighborhoodIterator.

Adapted ShapedImageNeighborhoodRange to allow the user to specify an optional pixel access parameter. This parameter allows specifying the constant value for this policy. A policy class that supports this pixel access parameter must have a nested PixelAccessParameterType type.

Change-Id: I5a1ae9f918a31d03a4dd3d84ceedd4385ab4ec7d