ENH: Added GradientNormThreshold to HoughTransform2DCirclesImageFilter
Allowed the user to specify when the gradient is considered
significantly greater than a flat gradient, within
HoughTransform2DCirclesImageFilter::GenerateData().
This is relevant because pixels whose gradient is considered "flat"
are skipped by the filter.

The results of old user code might be slightly different than before,
because the check if the gradient is not flat has basically changed
from 'if ((fabs(Vx) > 1) || (fabs(Vy) > 1))' to
'if (sqrt(Vx*Vx + Vy*Vy) > 1)', when using the default value of
GradientNormThreshold. However, no large differences are expected.

Added VerifyPreconditions() to verify GradientNormThreshold >= 0.

Change-Id: I984314014b50d5d2d315e2933db942cf4eb6fec6