ENH: Add HalfHermitianFrequencyIterator
Add the missing FrequencyIterator dealing with images
in the frequency domain that have been generated from
RealToHalfHermitianForwardFFTImageFilter.

The layout of these images changes in the X Dimension (fastest).
If the original image in the spatial domain has (N0, N1) elements.
The output has: (N0/2 + 1 , N1).

The fact that N0/2 is an integer division make the inverse transform,
and the frequency iterator in this case, to require extra information
about the original size of the image.

Following the same name used in HalfHermitianToRealInverseFFT and
others, we provide ALL the frequency iterators with the member

`bool ActualXDimensionIsOdd`

Only affects in the computation of FrequencySpacing (this affects
GetFrequencyBin) in the case of the newly introduced
itkFrequencyHalfHermitianFFTLayoutImageRegionConstIteratorWithIndex.

In the other iterators is there to provide a common interface, but does
nothing.

Change-Id: Icc376ccbdc0f5ff539c5504234b6e07a4c85149b